+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Version 2.8.4
   * Allow input of Vp/Vs ratio under File->Options
+  * Add option to save configuration file through File menu
  
 ## Version 2.8.3
   * Update seisFile library to 1.8.0